requirements/base.txt
protobuf
grpcio
grpcio-tools
grpcio-reflection
psycopg2-binary
python-ulid
pyyaml
pydantic
